feat: PlaceCard 분리 (#4)
